fix: correct TupleGetter descriptor protocol argument indexing
The __get__ method in TupleGetter was expecting self as args[0], but
since it's implemented as a closure that captures self (tg), the
arguments should be:
- args[0]: instance (the namedtuple instance)
- args[1]: owner (the namedtuple class, optional)

Changed from expecting 2 args with args[1] as instance to expecting 1 arg
with args[0] as instance.

This fixes namedtuple field access - now p.x returns the actual value (10)
instead of the tuplegetter descriptor object.