fix: make iter() raise TypeError instead of generic Exception
When iter() is called on a non-iterable object, it should raise a
TypeError that can be caught by Python code. Previously it was raising
a generic fmt.Errorf() which couldn't be caught by except TypeError.

This fixes unittest's _isnotsuite() function which relies on catching
TypeError to distinguish test cases from test suites.

Changes:
- builtin/iteration_protocol.go: Return &core.TypeError{} instead of
  fmt.Errorf() when object is not iterable